+++ src/factory/Factory.js
@@ -13,14 +13,24 @@ Factory.prototype = {
 	register : function(type,param,services){
 		type = type.toLowerCase();
 
-		var clazz = new Function("clazz","var args=[];for(var i=1;i<arguments.length;i++)args.push(arguments[i]);clazz.apply(this,args)");
-
-		var props = {};
-		
 		var data = param.data;
 		delete param.data;
 
-		Util.ext(props,param);
+		var props = {};
+		
+		if(typeof param == 'string'){
+			data = param;
+		}else{
+			Util.ext(props,param);
+		}
+
+		//re register
+		if(this.types[type]){
+			this.types[type].data = data;
+			this.types[type].props = props;
+			return;
+		}
+		var clazz = new Function("clazz","var args=[];for(var i=1;i<arguments.length;i++)args.push(arguments[i]);clazz.apply(this,args)");
 		Util.inherits(clazz,this.baseClass);
 
 		this.types[type] = {clazz:clazz,props:props,services:services,data:data};